Changed name from components to modules. Extracted model
