+++ src/modules/auth/auth.routes.ts
@@ -1,5 +1,5 @@
-import { NextFunction, Request, Response } from 'express';
-import { Router } from 'express';
+import { NextFunction, Request, Response, Router } from 'express';
+
 import AuthValidation from './auth.validation';
 import AuthController from './auth.controller';
 import { isAuthenticated } from '@middlewares';
@@ -20,11 +20,11 @@ router.post('/refresh', AuthController.refresh);
 
 // GET /secured - for testing purposes
 router.get('/secured', isAuthenticated, (req: Request, res: Response, next: NextFunction) => {
-    try {
-        res.json('ok!');
-    } catch (err) {
-        next(err);
-    }
+  try {
+    res.json('ok!');
+  } catch (err) {
+    next(err);
+  }
 });
 
 export { router as authRoutes };
