+++ src/modules/auth/auth.validation.ts
@@ -1,4 +1,5 @@
 import { body } from 'express-validator';
+
 import { User } from './auth.model';
 import AuthService from './auth.service';
 
@@ -10,13 +11,13 @@ export default [
     .isEmail()
     .withMessage('Please provide correct email address.')
     .custom(async (email: string) => {
-        const isEmailExists = await AuthService.isEmailExists(email);
+      const isEmailExists = await AuthService.isEmailExists(email);
 
-        if (isEmailExists) {
-            throw new Error('This e-mail address already exists.');
-        }
+      if (isEmailExists) {
+        throw new Error('This e-mail address already exists.');
+      }
     }),
   body('password' as keyof User)
     .notEmpty()
-    .withMessage('Password should not be empty.')
+    .withMessage('Password should not be empty.'),
 ];
