+++ src/modules/image/image.controller.ts
