+++ src/modules/image/image.routes.ts
