+++ tsconfig.json
@@ -1,19 +1,17 @@
 {
   "compilerOptions": {
-    "typeRoots": ["src/@types"],
     "target": "es2016",
     "module": "commonjs",
     "rootDir": ".",
     "moduleResolution": "node",
     "baseUrl": ".",
     "paths": {
-      "@components": ["src/components"],
-      "@config": ["src/config"],
+      "@modules": ["src/modules"],
       "@middlewares": ["src/middlewares"],
-      "@services/*": ["src/services/*"],
       "@utils": ["src/utils"],
       "@utils/*": ["src/utils/*"],
-      "@db/*": ["src/db/*"]
+      "@db/*": ["src/db/*"],
+      "@models": ["src/models"]
     },
     "resolveJsonModule": true,
     "sourceMap": true,
@@ -22,13 +20,9 @@
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
-    "skipLibCheck": true,
+    "skipLibCheck": true
   },
   "ts-node": {
     "require": ["tsconfig-paths/register"]
-  },
-  "include": [
-    "src/**/*.ts",
-    "src/@types/**/*.d.ts"
-  ]
+  }
 }
